+++ bricks/flow-builder/src/schema-editor/components/ref-required-item/RefRequiredItem.tsx
@@ -4,16 +4,18 @@ import { useCurModel } from "../../hooks/useCurModel";
 
 export interface RefRequiredItem {
   value?: string[];
+  prefix?: string;
   onChange?: (value: string[]) => void;
   model: string;
 }
 
 export function RefRequiredItem(props: RefRequiredItem): React.ReactElement {
-  const [{ modelData }, setModelName] = useCurModel(props.model);
+  const { prefix, model, value } = props;
+  const [{ modelData }, setModelName] = useCurModel(model);
 
   useEffect(() => {
-    setModelName(props.model);
-  }, [props.model, setModelName]);
+    setModelName(model);
+  }, [model, setModelName]);
 
   const handleChange = (value: string[]): void => {
     props.onChange?.(value);
@@ -21,9 +23,12 @@ export function RefRequiredItem(props: RefRequiredItem): React.ReactElement {
 
   return (
     <div>
-      <Select mode="tags" value={props.value} onChange={handleChange}>
+      <Select mode="tags" value={value} onChange={handleChange}>
         {modelData.fields?.map((item) => (
-          <Select.Option key={item.name} value={`${props.model}.${item.name}`}>
+          <Select.Option
+            key={item.name}
+            value={`${prefix ? prefix + "." : ""}${model}.${item.name}`}
+          >
             {item.name}
           </Select.Option>
         ))}