+++ bricks/forms/src/cmdb-cascader/index.spec.ts
@@ -26,6 +26,17 @@ describe("forms.cmdb-cascader", () => {
     await jest.runAllTimers();
     element.addEventListener("forms.cmdb-cascader.change", fn);
     element.handleChange("xxx");
-    expect(fn).toBeCalled();
+    expect(fn).toBeCalledTimes(0);
+  });
+
+  it("handleChangeV2 event", async () => {
+    const fn = jest.fn();
+    const element = document.createElement("forms.cmdb-cascader");
+    // Always waiting for async `(dis)connectedCallback`
+    document.body.appendChild(element);
+    await jest.runAllTimers();
+    element.addEventListener("forms.cmdb-cascader.change.v2", fn);
+    element.handleChangeV2("xxx");
+    expect(fn).toBeCalledTimes(1);
   });
 });