+++ bricks/forms/src/dynamic-form-item-v2/autoCompleteItem.tsx
@@ -6,7 +6,8 @@ interface AutoCompleteProps {
   options: string[] | OptionType[];
   disabled?: boolean;
   allowClear?: boolean;
-  placeholder: React.ReactNode;
+  placeholder?: React.ReactNode;
+  isAppendMode?: boolean;
   value?: string;
   onChange?: (value: string) => void;
 }
@@ -32,7 +33,8 @@ function filterOptions(value: string, options: OptionType[]): OptionType[] {
 }
 
 export function AutoCompleteItem(props: AutoCompleteProps): React.ReactElement {
-  const { disabled, allowClear, placeholder, value, onChange } = props;
+  const { disabled, allowClear, placeholder, isAppendMode, value, onChange } =
+    props;
 
   const originalOptions: OptionType[] = useMemo(
     () =>
@@ -45,6 +47,8 @@ export function AutoCompleteItem(props: AutoCompleteProps): React.ReactElement {
     [props.options]
   );
   const [options, setOptions] = React.useState(originalOptions);
+  const [nodeId, setNodeId] = React.useState("");
+  const [node, setNode] = React.useState<HTMLInputElement>();
 
   const onSearch = (v: string) => {
     const q = v.trim().toLowerCase();
@@ -56,6 +60,36 @@ export function AutoCompleteItem(props: AutoCompleteProps): React.ReactElement {
     setOptions(originalOptions);
   }, [originalOptions]);
 
+  React.useEffect(() => {
+    setNode(document.getElementById(nodeId) as HTMLInputElement);
+  }, [nodeId]);
+
+  const handleAppendChange = (e: string) => {
+    if (e) {
+      if (
+        e === value ||
+        value === undefined ||
+        Math.abs(e?.length - value?.length) >= 1
+      ) {
+        onChange(e);
+      } else {
+        onChange(value.concat(e));
+      }
+    } else {
+      onChange("");
+    }
+  };
+
+  const handleAppendSelect = (e: string) => {
+    if (!value || !e) return;
+    onChange(
+      value
+        .slice(0, node?.selectionStart ?? value.length)
+        .concat(e)
+        .concat(value.slice(node?.selectionStart ?? value.length))
+    );
+  };
+
   return (
     <AutoComplete
       options={options}
@@ -63,8 +97,10 @@ export function AutoCompleteItem(props: AutoCompleteProps): React.ReactElement {
       allowClear={allowClear}
       placeholder={placeholder}
       value={value}
-      onSearch={onSearch}
-      onChange={onChange}
+      onSelect={isAppendMode ? handleAppendSelect : null}
+      onSearch={isAppendMode ? null : onSearch}
+      onChange={isAppendMode ? handleAppendChange : onChange}
+      onFocus={(e) => setNodeId(e.target.id)}
     />
   );
 }