+++ bricks/forms/src/general-radio/index.spec.ts
@@ -1,7 +1,9 @@
 import ReactDOM from "react-dom";
 import "./";
 
-const spyOnRender = jest.spyOn(ReactDOM, "render").mockImplementation(() => null);
+const spyOnRender = jest
+  .spyOn(ReactDOM, "render")
+  .mockImplementation(() => null);
 const unmountComponentAtNode = jest
   .spyOn(ReactDOM, "unmountComponentAtNode")
   .mockImplementation((() => null) as any);
@@ -18,15 +20,23 @@ describe("forms.general-radio", () => {
 
     const dispatchEvent = jest.spyOn(element, "dispatchEvent");
 
-    const props = spyOnRender.mock.calls[spyOnRender.mock.calls.length - 1][0].props
-    .children.props;
-    const value = ['a'];
+    const props =
+      spyOnRender.mock.calls[spyOnRender.mock.calls.length - 1][0].props
+        .children.props;
+    const value = ["a"];
     props.onChange(value);
     await (global as any).flushPromises();
     expect(dispatchEvent).lastCalledWith(
       expect.objectContaining({ type: "general.radio.change", detail: value })
     );
-
+    props.onChangeV2(value);
+    await (global as any).flushPromises();
+    expect(dispatchEvent).lastCalledWith(
+      expect.objectContaining({
+        type: "general.radio.change.v2",
+        detail: value,
+      })
+    );
     document.body.removeChild(element);
     await jest.runAllTimers();
     expect(unmountComponentAtNode).toBeCalled();