+++ bricks/frame-bricks/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.2](https://github.com/easyops-cn/next-basics/compare/@next-bricks/frame-bricks@1.9.1...@next-bricks/frame-bricks@1.9.2) (2023-02-03)
+
+
+### Bug Fixes
+
+* 修改sidebar的border-radius ([a3f2162](https://github.com/easyops-cn/next-basics/commit/a3f2162837bdc5cbf006870bf8c635c543cda3f4))
+
+
+
+
+
 ## [1.9.1](https://github.com/easyops-cn/next-basics/compare/@next-bricks/frame-bricks@1.9.0...@next-bricks/frame-bricks@1.9.1) (2023-01-11)
 
 **Note:** Version bump only for package @next-bricks/frame-bricks